refactor: add a structure to the code
