+++ argocue/helm/helm.go
@@ -1,4 +1,4 @@
-package argocue
+package helm
 
 import (
 	"encoding/json"
@@ -9,6 +9,7 @@ import (
 	"strings"
 
 	"github.com/darklab8/argocd-cue/argocue/logus"
+	"github.com/darklab8/argocd-cue/argocue/types"
 	"github.com/darklab8/go-typelog/typelog"
 	"github.com/darklab8/go-utils/goutils/utils/utils_filepath"
 	"github.com/darklab8/go-utils/goutils/utils/utils_types"
@@ -24,14 +25,14 @@ func helmLoadDeps(workdir utils_types.FilePath) {
 	build := exec.Command("helm", "dep", "update", "--kube-insecure-skip-tls-verify")
 	build.Dir = workdir.ToString()
 	build_out, err := build.Output()
-	HandleCmdError(build_out, err, "failed to run helm dep update")
+	types.HandleCmdError(build_out, err, "failed to run helm dep update")
 }
 
 func buildHelm(workdir utils_types.FilePath) {
 	build := exec.Command("cue", "cmd", "build")
 	build.Dir = workdir.ToString()
 	build_out, err := build.Output()
-	HandleCmdError(build_out, err, "failed to cue cmd build")
+	types.HandleCmdError(build_out, err, "failed to cue cmd build")
 }
 
 type ParameterGroup struct {
@@ -86,12 +87,12 @@ func (h Helm) Generate(workdir utils_types.FilePath) {
 	cmd := exec.Command(command_exec, templating_commands...)
 	cmd.Dir = workdir.ToString()
 	out, err := cmd.Output()
-	HandleCmdError(out, err, "failed to helm template")
+	types.HandleCmdError(out, err, "failed to helm template")
 	fmt.Println(string(out))
 }
 
-func newHelmParams(Map map[string]interface{}) []ApplicationParams {
-	return []ApplicationParams{
+func newHelmParams(Map map[string]interface{}) []types.ApplicationParams {
+	return []types.ApplicationParams{
 		{
 			Name:           "helm-parameters",
 			Title:          "Helm Parameters",
