+++ argocue/manifests/manifests.go
@@ -1,4 +1,4 @@
-package argocue
+package manifests
 
 import (
 	"encoding/json"
@@ -6,6 +6,7 @@ import (
 	"os/exec"
 
 	"github.com/darklab8/argocd-cue/argocue/logus"
+	"github.com/darklab8/argocd-cue/argocue/types"
 	"github.com/darklab8/go-utils/goutils/utils/utils_types"
 )
 
@@ -19,12 +20,12 @@ func (m Manifests) Generate(workdir utils_types.FilePath) {
 	cmd.Dir = workdir.ToString()
 	out, err := cmd.Output()
 
-	HandleCmdError(out, err, "failed to run cue cmd dump")
+	types.HandleCmdError(out, err, "failed to run cue cmd dump")
 	fmt.Println(string(out))
 }
 
-func mewManifestsParams(Map map[string]interface{}) []ApplicationParams {
-	return []ApplicationParams{
+func mewManifestsParams(Map map[string]interface{}) []types.ApplicationParams {
+	return []types.ApplicationParams{
 		{
 			Name:           "manifests-parameters",
 			Title:          "Manifests Parameters",
