+++ argocue/manifests/manifests_test.go
@@ -1,21 +1,19 @@
-package argocue
+package manifests
 
 import (
 	"testing"
 
-	"github.com/darklab8/go-utils/goutils/utils"
+	"github.com/darklab8/argocd-cue/argocue/types"
 	"github.com/darklab8/go-utils/goutils/utils/utils_filepath"
 	"github.com/darklab8/go-utils/goutils/utils/utils_types"
 )
 
-var ProjectRoot = utils_filepath.Dir(utils.GetCurrentFolder())
-
 func TestManifests(t *testing.T) {
-	manifests_folder := utils_filepath.Join(ProjectRoot, "examples", "manifests")
+	manifests_folder := utils_filepath.Join(types.ProjectRoot, "examples", "manifests")
 	NewManifests().Generate(utils_types.FilePath(manifests_folder))
 }
 
 func TestManifestsParams(t *testing.T) {
-	manifests_folder := utils_filepath.Join(ProjectRoot, "examples", "manifests")
+	manifests_folder := utils_filepath.Join(types.ProjectRoot, "examples", "manifests")
 	NewManifests().GetParameters(utils_types.FilePath(manifests_folder))
 }
